refactor: Move mindthegap cmd to own package
This enables other standalone binaries to be created.
