+++ cmd/mindthegap/create/create.go
@@ -1,22 +1,24 @@
 // Copyright 2021 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package importcmd
+package create
 
 import (
 	"github.com/spf13/cobra"
 
 	"github.com/mesosphere/dkp-cli-runtime/core/output"
 
-	"github.com/mesosphere/mindthegap/cmd/importcmd/imagebundle"
+	"github.com/mesosphere/mindthegap/cmd/mindthegap/create/helmbundle"
+	"github.com/mesosphere/mindthegap/cmd/mindthegap/create/imagebundle"
 )
 
 func NewCommand(out output.Output) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "import",
-		Short: "Import images from an image bundle into Containerd",
+		Use:   "create",
+		Short: "Create an tar.gz image or helm bundle",
 	}
 
 	cmd.AddCommand(imagebundle.NewCommand(out))
+	cmd.AddCommand(helmbundle.NewCommand(out))
 	return cmd
 }
