+++ cmd/mindthegap/importcmd/import.go
@@ -1,24 +1,22 @@
 // Copyright 2021 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package create
+package importcmd
 
 import (
 	"github.com/spf13/cobra"
 
 	"github.com/mesosphere/dkp-cli-runtime/core/output"
 
-	"github.com/mesosphere/mindthegap/cmd/create/helmbundle"
-	"github.com/mesosphere/mindthegap/cmd/create/imagebundle"
+	"github.com/mesosphere/mindthegap/cmd/mindthegap/importcmd/imagebundle"
 )
 
 func NewCommand(out output.Output) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "create",
-		Short: "Create an tar.gz image or helm bundle",
+		Use:   "import",
+		Short: "Import images from an image bundle into Containerd",
 	}
 
 	cmd.AddCommand(imagebundle.NewCommand(out))
-	cmd.AddCommand(helmbundle.NewCommand(out))
 	return cmd
 }
