+++ cmd/mindthegap/serve/imagebundle/image_bundle.go
@@ -13,7 +13,7 @@ import (
 	"github.com/mesosphere/dkp-cli-runtime/core/output"
 
 	"github.com/mesosphere/mindthegap/cleanup"
-	"github.com/mesosphere/mindthegap/cmd/utils"
+	"github.com/mesosphere/mindthegap/cmd/mindthegap/utils"
 	"github.com/mesosphere/mindthegap/config"
 	"github.com/mesosphere/mindthegap/docker/registry"
 )
